modify : user, auth 모듈 라우팅 수정 및 컨트롤러 분리
