+++ src/swagger/auth/Refresh.decorator.ts
@@ -1,7 +1,7 @@
 import {
   RefreshOkDto,
   RefreshUnauthorizedDto,
-} from './../../auth/dto/auth-res.dto';
+} from '../../user/application/dtos/auth-res.dto';
 import { applyDecorators } from '@nestjs/common';
 import {
   ApiCookieAuth,
