+++ src/swagger/auth/SignUp.decorator.ts
@@ -1,7 +1,7 @@
 import {
   AuthSignUpConflictDto,
   AuthSignUpCreatedDto,
-} from './../../auth/dto/auth-res.dto';
+} from '../../user/application/dtos/auth-res.dto';
 import { applyDecorators } from '@nestjs/common';
 import {
   ApiConflictResponse,
