+++ src/swagger/auth/SingOut.decorator.ts
@@ -1,7 +1,7 @@
 import {
   AuthSignOutUnauthorizedDto,
   SignOutOkDto,
-} from '../../auth/dto/auth-res.dto';
+} from '../../user/application/dtos/auth-res.dto';
 import { applyDecorators } from '@nestjs/common';
 import {
   ApiCookieAuth,
