+++ src/user/application/commands/update-user/update-user.request.dto.ts
@@ -1,0 +1,15 @@
+import { IsBoolean, IsNotEmpty, IsNumber } from 'class-validator';
+
+export class UpdateUserDto {
+  @IsNotEmpty()
+  @IsNumber()
+  readonly fieldId: number;
+
+  @IsNotEmpty()
+  @IsNumber()
+  readonly careerId: number;
+
+  @IsNotEmpty()
+  @IsBoolean()
+  readonly isKorean: boolean;
+}
